// src/Components/Types/DataType.tsx
export type Data = {
    teamId1: number;
    teamId2: number;
    teamName1: string;
    teamName2: string;
    shortName1: string;
    shortName2: string;
    teamIconUrl1: string;
    teamIconUrl2: string;
    matchIsFinished1: boolean;
    matchIsFinished2: boolean;
    matchResults1: Array<any>;
    matchResults2: Array<any>;
};